requirements.txt
bcrypt==4.0.0
blinker==1.4
certifi==2020.6.20
cffi==1.11.5
click==8.0.3
dateutils==0.6.12
Flask==2.1.2
Flask_Bcrypt==1.0.1
Flask_Login==0.6.2
Flask_SQLAlchemy==2.5.1
Flask_WTF==1.0.1
WTForms==3.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Pillow==9.2.0
six==1.16.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2